[ML] Data Frame Analytics: Expandable sections for classification and regression (#79414)
Applies the expandable section based layout to the results pages of classification and regression analytics jobs.
