+++ x-pack/plugins/ml/public/application/data_frame_analytics/pages/analytics_exploration/components/classification_exploration/_classification_exploration.scss
@@ -23,7 +23,7 @@
 .mlDataFrameAnalyticsClassification__actualLabel {
   float: left;
   width: 80px;
-  padding-top: $euiSize * 4 + $euiSizeXS;
+  padding-top: $euiSize * 4;
 }
 
 /*
