+++ x-pack/plugins/ml/public/application/data_frame_analytics/pages/analytics_exploration/components/classification_exploration/classification_exploration.tsx
@@ -17,17 +17,19 @@ interface Props {
 }
 
 export const ClassificationExploration: FC<Props> = ({ jobId, defaultIsTraining }) => (
-  <ExplorationPageWrapper
-    jobId={jobId}
-    title={i18n.translate(
-      'xpack.ml.dataframe.analytics.classificationExploration.tableJobIdTitle',
-      {
-        defaultMessage: 'Destination index for classification job ID {jobId}',
-        values: { jobId },
-      }
-    )}
-    EvaluatePanel={EvaluatePanel}
-    FeatureImportanceSummaryPanel={FeatureImportanceSummaryPanel}
-    defaultIsTraining={defaultIsTraining}
-  />
+  <div className="mlDataFrameAnalyticsClassification">
+    <ExplorationPageWrapper
+      jobId={jobId}
+      title={i18n.translate(
+        'xpack.ml.dataframe.analytics.classificationExploration.tableJobIdTitle',
+        {
+          defaultMessage: 'Destination index for classification job ID {jobId}',
+          values: { jobId },
+        }
+      )}
+      EvaluatePanel={EvaluatePanel}
+      FeatureImportanceSummaryPanel={FeatureImportanceSummaryPanel}
+      defaultIsTraining={defaultIsTraining}
+    />
+  </div>
 );
