+++ x-pack/plugins/ml/public/application/data_frame_analytics/pages/analytics_exploration/components/classification_exploration/evaluate_panel.tsx
@@ -6,7 +6,7 @@
 
 import './_classification_exploration.scss';
 
-import React, { FC, useState, useEffect, Fragment } from 'react';
+import React, { FC, useState, useEffect } from 'react';
 import { i18n } from '@kbn/i18n';
 import { FormattedMessage } from '@kbn/i18n/react';
 import {
@@ -15,7 +15,6 @@ import {
   EuiFlexGroup,
   EuiFlexItem,
   EuiIconTip,
-  EuiPanel,
   EuiSpacer,
   EuiText,
   EuiTitle,
@@ -30,7 +29,6 @@ import {
   DataFrameAnalyticsConfig,
 } from '../../../../common';
 import { isKeywordAndTextType } from '../../../../common/fields';
-import { getTaskStateBadge } from '../../../analytics_management/components/analytics_list/use_columns';
 import { DATA_FRAME_TASK_STATE } from '../../../analytics_management/components/analytics_list/common';
 import {
   isResultsSearchBoolQuery,
@@ -39,7 +37,9 @@ import {
   ResultsSearchQuery,
   ANALYSIS_CONFIG_TYPE,
 } from '../../../../common/analytics';
-import { LoadingPanel } from '../loading_panel';
+
+import { ExpandableSection, HEADER_ITEMS_LOADING } from '../expandable_section';
+
 import {
   getColumnData,
   ACTUAL_CLASS_ID,
@@ -47,7 +47,7 @@ import {
   getTrailingControlColumns,
 } from './column_data';
 
-interface Props {
+export interface EvaluatePanelProps {
   jobConfig: DataFrameAnalyticsConfig;
   jobStatus?: DATA_FRAME_TASK_STATE;
   searchQuery: ResultsSearchQuery;
@@ -90,7 +90,7 @@ function getHelpText(dataSubsetTitle: string) {
   return helpText;
 }
 
-export const EvaluatePanel: FC<Props> = ({ jobConfig, jobStatus, searchQuery }) => {
+export const EvaluatePanel: FC<EvaluatePanelProps> = ({ jobConfig, jobStatus, searchQuery }) => {
   const {
     services: { docLinks },
   } = useMlKibana();
@@ -272,10 +272,6 @@ export const EvaluatePanel: FC<Props> = ({ jobConfig, jobStatus, searchQuery }) 
     return <span>{columnId === ACTUAL_CLASS_ID ? cellValue : accuracy}</span>;
   };
 
-  if (isLoading === true) {
-    return <LoadingPanel />;
-  }
-
   const { ELASTIC_WEBSITE_URL, DOC_LINK_VERSION } = docLinks;
 
   const showTrailingColumns = columnsData.length > MAX_COLUMNS;
@@ -288,137 +284,159 @@ export const EvaluatePanel: FC<Props> = ({ jobConfig, jobStatus, searchQuery }) 
     showTrailingColumns === true && showFullColumns === false ? MAX_COLUMNS : columnsData.length;
 
   return (
-    <EuiPanel
-      data-test-subj="mlDFAnalyticsClassificationExplorationEvaluatePanel"
-      className="mlDataFrameAnalyticsClassification"
-    >
-      <div>
-        <EuiFlexGroup alignItems="center" justifyContent="spaceBetween">
-          <EuiFlexItem grow={false}>
-            <EuiTitle size="xs">
-              <span>
-                {i18n.translate(
-                  'xpack.ml.dataframe.analytics.classificationExploration.evaluateJobIdTitle',
-                  {
-                    defaultMessage: 'Evaluation of classification job ID {jobId}',
-                    values: { jobId: jobConfig.id },
-                  }
-                )}
-              </span>
-            </EuiTitle>
-          </EuiFlexItem>
-          {jobStatus !== undefined && (
-            <EuiFlexItem grow={false}>
-              <span>{getTaskStateBadge(jobStatus)}</span>
-            </EuiFlexItem>
-          )}
-          <EuiFlexItem />
-          <EuiFlexItem grow={false}>
-            <EuiButtonEmpty
-              target="_blank"
-              iconType="help"
-              iconSide="left"
-              color="primary"
-              href={`${ELASTIC_WEBSITE_URL}guide/en/machine-learning/${DOC_LINK_VERSION}/ml-dfanalytics-evaluate.html#ml-dfanalytics-classification`}
-            >
-              {i18n.translate(
-                'xpack.ml.dataframe.analytics.classificationExploration.classificationDocsLink',
-                {
-                  defaultMessage: 'Classification evaluation docs ',
-                }
-              )}
-            </EuiButtonEmpty>
-          </EuiFlexItem>
-        </EuiFlexGroup>
-      </div>
-      {error !== null && <ErrorCallout error={error} />}
-      {error === null && (
-        <Fragment>
-          <div>
-            <EuiFlexGroup gutterSize="xs">
-              <EuiTitle size="xxs">
-                <span>{getHelpText(dataSubsetTitle)}</span>
-              </EuiTitle>
-              <EuiFlexItem grow={false}>
-                <EuiIconTip
-                  anchorClassName="mlDataFrameAnalyticsClassificationInfoTooltip"
-                  content={i18n.translate(
-                    'xpack.ml.dataframe.analytics.classificationExploration.confusionMatrixTooltip',
-                    {
-                      defaultMessage:
-                        'The multi-class confusion matrix contains the number of occurrences where the analysis classified data points correctly with their actual class as well as the number of occurrences where it misclassified them with another class',
-                    }
-                  )}
-                />
-              </EuiFlexItem>
-            </EuiFlexGroup>
-          </div>
-          {docsCount !== null && (
-            <EuiText size="xs" color="subdued">
-              <FormattedMessage
-                id="xpack.ml.dataframe.analytics.classificationExploration.generalizationDocsCount"
-                defaultMessage="{docsCount, plural, one {# doc} other {# docs}} evaluated"
-                values={{ docsCount }}
-              />
-            </EuiText>
-          )}
-          {/* BEGIN TABLE ELEMENTS */}
-          <EuiSpacer size="m" />
-          <div className="mlDataFrameAnalyticsClassification__confusionMatrix">
-            <div className="mlDataFrameAnalyticsClassification__actualLabel">
-              <EuiText size="xs" color="subdued">
-                <FormattedMessage
-                  id="xpack.ml.dataframe.analytics.classificationExploration.confusionMatrixActualLabel"
-                  defaultMessage="Actual label"
-                />
-              </EuiText>
-            </div>
-            <div className="mlDataFrameAnalyticsClassification__dataGridMinWidth">
-              {columns.length > 0 && columnsData.length > 0 && (
-                <>
-                  <div>
-                    <EuiText size="xs" color="subdued">
-                      <FormattedMessage
-                        id="xpack.ml.dataframe.analytics.classificationExploration.confusionMatrixPredictedLabel"
-                        defaultMessage="Predicted label"
-                      />
-                    </EuiText>
-                  </div>
-                  <EuiSpacer size="s" />
-                  <EuiDataGrid
-                    data-test-subj="mlDFAnalyticsClassificationExplorationConfusionMatrix"
-                    aria-label={i18n.translate(
-                      'xpack.ml.dataframe.analytics.classificationExploration.confusionMatrixLabel',
+    <>
+      <ExpandableSection
+        dataTestId="ClassificationEvaluation"
+        title={
+          <FormattedMessage
+            id="xpack.ml.dataframe.analytics.classificationExploration.evaluateSectionTitle"
+            defaultMessage="Model evaluation"
+          />
+        }
+        docsLink={
+          <EuiButtonEmpty
+            target="_blank"
+            iconType="help"
+            iconSide="left"
+            color="primary"
+            href={`${ELASTIC_WEBSITE_URL}guide/en/machine-learning/${DOC_LINK_VERSION}/ml-dfanalytics-evaluate.html#ml-dfanalytics-classification`}
+          >
+            {i18n.translate(
+              'xpack.ml.dataframe.analytics.classificationExploration.classificationDocsLink',
+              {
+                defaultMessage: 'Classification evaluation docs ',
+              }
+            )}
+          </EuiButtonEmpty>
+        }
+        headerItems={
+          !isLoading
+            ? [
+                ...(jobStatus !== undefined
+                  ? [
                       {
-                        defaultMessage: 'Classification confusion matrix',
-                      }
-                    )}
-                    columns={shownColumns}
-                    columnVisibility={{ visibleColumns, setVisibleColumns }}
-                    rowCount={rowCount}
-                    renderCellValue={renderCellValue}
-                    inMemory={{ level: 'sorting' }}
-                    toolbarVisibility={{
-                      showColumnSelector: true,
-                      showStyleSelector: false,
-                      showFullScreenSelector: false,
-                      showSortSelector: false,
-                    }}
-                    popoverContents={popoverContents}
-                    gridStyle={{ rowHover: 'none' }}
-                    trailingControlColumns={
-                      showTrailingColumns === true && showFullColumns === false
-                        ? getTrailingControlColumns(extraColumns, setShowFullColumns)
-                        : undefined
-                    }
-                  />
+                        id: 'jobStatus',
+                        label: i18n.translate(
+                          'xpack.ml.dataframe.analytics.classificationExploration.evaluateJobStatusLabel',
+                          {
+                            defaultMessage: 'Job status',
+                          }
+                        ),
+                        value: jobStatus,
+                      },
+                    ]
+                  : []),
+                ...(docsCount !== null
+                  ? [
+                      {
+                        id: 'docsEvaluated',
+                        label: i18n.translate(
+                          'xpack.ml.dataframe.analytics.classificationExploration.generalizationDocsCount',
+                          {
+                            defaultMessage: '{docsCount, plural, one {doc} other {docs}} evaluated',
+                            values: { docsCount },
+                          }
+                        ),
+                        value: docsCount,
+                      },
+                    ]
+                  : []),
+              ]
+            : HEADER_ITEMS_LOADING
+        }
+        contentPadding={true}
+        content={
+          !isLoading ? (
+            <>
+              {error !== null && <ErrorCallout error={error} />}
+              {error === null && (
+                <>
+                  <EuiFlexGroup gutterSize="none">
+                    <EuiTitle size="xxs">
+                      <span>{getHelpText(dataSubsetTitle)}</span>
+                    </EuiTitle>
+                    <EuiFlexItem grow={false}>
+                      <EuiIconTip
+                        anchorClassName="mlDataFrameAnalyticsClassificationInfoTooltip"
+                        content={i18n.translate(
+                          'xpack.ml.dataframe.analytics.classificationExploration.confusionMatrixTooltip',
+                          {
+                            defaultMessage:
+                              'The multi-class confusion matrix contains the number of occurrences where the analysis classified data points correctly with their actual class as well as the number of occurrences where it misclassified them with another class',
+                          }
+                        )}
+                      />
+                    </EuiFlexItem>
+                  </EuiFlexGroup>
+                  {/* BEGIN TABLE ELEMENTS */}
+                  <EuiSpacer size="m" />
+                  <div className="mlDataFrameAnalyticsClassification__confusionMatrix">
+                    <div className="mlDataFrameAnalyticsClassification__actualLabel">
+                      <EuiText size="xs" color="subdued">
+                        <FormattedMessage
+                          id="xpack.ml.dataframe.analytics.classificationExploration.confusionMatrixActualLabel"
+                          defaultMessage="Actual label"
+                        />
+                      </EuiText>
+                    </div>
+                    <div className="mlDataFrameAnalyticsClassification__dataGridMinWidth">
+                      {columns.length > 0 && columnsData.length > 0 && (
+                        <>
+                          <div>
+                            <EuiText size="xs" color="subdued">
+                              <FormattedMessage
+                                id="xpack.ml.dataframe.analytics.classificationExploration.confusionMatrixPredictedLabel"
+                                defaultMessage="Predicted label"
+                              />
+                            </EuiText>
+                          </div>
+                          <EuiSpacer size="s" />
+                          <EuiDataGrid
+                            data-test-subj="mlDFAnalyticsClassificationExplorationConfusionMatrix"
+                            aria-label={i18n.translate(
+                              'xpack.ml.dataframe.analytics.classificationExploration.confusionMatrixLabel',
+                              {
+                                defaultMessage: 'Classification confusion matrix',
+                              }
+                            )}
+                            columns={shownColumns}
+                            columnVisibility={{ visibleColumns, setVisibleColumns }}
+                            rowCount={rowCount}
+                            renderCellValue={renderCellValue}
+                            inMemory={{ level: 'sorting' }}
+                            toolbarVisibility={{
+                              showColumnSelector: true,
+                              showStyleSelector: false,
+                              showFullScreenSelector: false,
+                              showSortSelector: false,
+                            }}
+                            popoverContents={popoverContents}
+                            gridStyle={{
+                              border: 'all',
+                              fontSize: 's',
+                              cellPadding: 's',
+                              stripes: false,
+                              rowHover: 'none',
+                              header: 'shade',
+                            }}
+                            trailingControlColumns={
+                              showTrailingColumns === true && showFullColumns === false
+                                ? getTrailingControlColumns(extraColumns, setShowFullColumns)
+                                : undefined
+                            }
+                          />
+                        </>
+                      )}
+                    </div>
+                  </div>
                 </>
               )}
-            </div>
-          </div>
-        </Fragment>
-      )}
-      {/* END TABLE ELEMENTS */}
-    </EuiPanel>
+              {/* END TABLE ELEMENTS */}
+            </>
+          ) : null
+        }
+      />
+      <EuiSpacer size="m" />
+    </>
   );
 };
