+++ x-pack/plugins/ml/public/application/data_frame_analytics/pages/analytics_exploration/components/expandable_section/expandable_section.scss
@@ -1,3 +1,7 @@
 .mlExpandableSection {
   padding: 0 $euiSizeS $euiSizeS $euiSizeS;
 }
+
+.mlExpandableSection-contentPadding {
+  padding: $euiSizeS;
+}
