+++ x-pack/plugins/ml/public/application/data_frame_analytics/pages/analytics_exploration/components/exploration_page_wrapper/exploration_page_wrapper.tsx
@@ -4,19 +4,48 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-import React, { FC, useState } from 'react';
+import React, { FC, useEffect, useState } from 'react';
 
-import { EuiSpacer } from '@elastic/eui';
+import { EuiFlexGroup, EuiFlexItem, EuiSpacer, EuiText } from '@elastic/eui';
+import { i18n } from '@kbn/i18n';
 
-import { useResultsViewConfig, DataFrameAnalyticsConfig } from '../../../../common';
-import { ResultsSearchQuery, defaultSearchQuery } from '../../../../common/analytics';
+import { useUrlState } from '../../../../../util/url_state';
+
+import {
+  defaultSearchQuery,
+  getDefaultTrainingFilterQuery,
+  useResultsViewConfig,
+  DataFrameAnalyticsConfig,
+} from '../../../../common';
+import { ResultsSearchQuery } from '../../../../common/analytics';
 
 import { DATA_FRAME_TASK_STATE } from '../../../analytics_management/components/analytics_list/common';
 
+import { ExpandableSectionAnalytics } from '../expandable_section';
 import { ExplorationResultsTable } from '../exploration_results_table';
+import { ExplorationQueryBar } from '../exploration_query_bar';
 import { JobConfigErrorCallout } from '../job_config_error_callout';
 import { LoadingPanel } from '../loading_panel';
 import { FeatureImportanceSummaryPanelProps } from '../total_feature_importance_summary/feature_importance_summary';
+
+const filters = {
+  options: [
+    {
+      id: 'training',
+      label: i18n.translate('xpack.ml.dataframe.analytics.explorationResults.trainingSubsetLabel', {
+        defaultMessage: 'Training',
+      }),
+    },
+    {
+      id: 'testing',
+      label: i18n.translate('xpack.ml.dataframe.analytics.explorationResults.testingSubsetLabel', {
+        defaultMessage: 'Testing',
+      }),
+    },
+  ],
+  columnId: 'ml.is_training',
+  key: { training: true, testing: false },
+};
 
 export interface EvaluatePanelProps {
   jobConfig: DataFrameAnalyticsConfig;
@@ -50,7 +79,25 @@ export const ExplorationPageWrapper: FC<Props> = ({
     needsDestIndexPattern,
     totalFeatureImportance,
   } = useResultsViewConfig(jobId);
+
   const [searchQuery, setSearchQuery] = useState<ResultsSearchQuery>(defaultSearchQuery);
+  const [globalState, setGlobalState] = useUrlState('_g');
+  const [defaultQueryString, setDefaultQueryString] = useState<string | undefined>();
+
+  useEffect(() => {
+    if (defaultIsTraining !== undefined && jobConfig !== undefined) {
+      // Apply defaultIsTraining filter
+      setSearchQuery(
+        getDefaultTrainingFilterQuery(jobConfig.dest.results_field, defaultIsTraining)
+      );
+      setDefaultQueryString(`${jobConfig.dest.results_field}.is_training : ${defaultIsTraining}`);
+      // Clear defaultIsTraining from url
+      setGlobalState('ml', {
+        analysisType: globalState.ml.analysisType,
+        jobId: globalState.ml.jobId,
+      });
+    }
+  }, [jobConfig?.dest.results_field]);
 
   if (jobConfigErrorMessage !== undefined || jobCapsServiceErrorMessage !== undefined) {
     return (
@@ -61,21 +108,54 @@ export const ExplorationPageWrapper: FC<Props> = ({
       />
     );
   }
+
   return (
     <>
+      {typeof jobConfig?.description !== 'undefined' && (
+        <>
+          <EuiText>{jobConfig?.description}</EuiText>
+          <EuiSpacer size="m" />
+        </>
+      )}
+
+      {indexPattern !== undefined && (
+        <>
+          <EuiFlexGroup direction="column">
+            <EuiFlexItem grow={false}>
+              <EuiSpacer size="s" />
+              <EuiFlexGroup justifyContent="spaceBetween">
+                <EuiFlexItem>
+                  <ExplorationQueryBar
+                    indexPattern={indexPattern}
+                    setSearchQuery={setSearchQuery}
+                    defaultQueryString={defaultQueryString}
+                    filters={filters}
+                  />
+                </EuiFlexItem>
+              </EuiFlexGroup>
+            </EuiFlexItem>
+          </EuiFlexGroup>
+          <EuiSpacer size="m" />
+        </>
+      )}
+
       {isLoadingJobConfig === true && jobConfig === undefined && <LoadingPanel />}
       {isLoadingJobConfig === false && jobConfig !== undefined && isInitialized === true && (
-        <EvaluatePanel jobConfig={jobConfig} jobStatus={jobStatus} searchQuery={searchQuery} />
+        <ExpandableSectionAnalytics jobId={jobConfig.id} />
       )}
+
       {isLoadingJobConfig === true && totalFeatureImportance === undefined && <LoadingPanel />}
       {isLoadingJobConfig === false && totalFeatureImportance !== undefined && (
         <>
-          <EuiSpacer />
           <FeatureImportanceSummaryPanel totalFeatureImportance={totalFeatureImportance} />
         </>
       )}
 
-      <EuiSpacer />
+      {isLoadingJobConfig === true && jobConfig === undefined && <LoadingPanel />}
+      {isLoadingJobConfig === false && jobConfig !== undefined && isInitialized === true && (
+        <EvaluatePanel jobConfig={jobConfig} jobStatus={jobStatus} searchQuery={searchQuery} />
+      )}
+
       {isLoadingJobConfig === true && jobConfig === undefined && <LoadingPanel />}
       {isLoadingJobConfig === false &&
         jobConfig !== undefined &&
@@ -86,9 +166,7 @@ export const ExplorationPageWrapper: FC<Props> = ({
             jobConfig={jobConfig}
             jobStatus={jobStatus}
             needsDestIndexPattern={needsDestIndexPattern}
-            setEvaluateSearchQuery={setSearchQuery}
-            title={title}
-            defaultIsTraining={defaultIsTraining}
+            searchQuery={searchQuery}
           />
         )}
     </>
