--- x-pack/plugins/ml/public/application/data_frame_analytics/pages/analytics_exploration/components/exploration_title/exploration_title.tsx
@@ -1,15 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License;
- * you may not use this file except in compliance with the Elastic License.
- */
-
-import React, { FC } from 'react';
-
-import { EuiTitle } from '@elastic/eui';
-
-export const ExplorationTitle: FC<{ title: string }> = ({ title }) => (
-  <EuiTitle size="xs">
-    <span>{title}</span>
-  </EuiTitle>
-);
