+++ x-pack/plugins/translations/translations/zh-CN.json
@@ -10566,7 +10566,6 @@
     "xpack.ml.dataframe.analytics.classificationExploration.confusionMatrixTestingHelpText": "用于测试数据集的标准化混淆矩阵",
     "xpack.ml.dataframe.analytics.classificationExploration.confusionMatrixTooltip": "多类混淆矩阵包含分析使用数据点的实际类正确分类数据点的次数以及分析使用其他类错误分类这些数据点的次数",
     "xpack.ml.dataframe.analytics.classificationExploration.confusionMatrixTrainingHelpText": "用于训练数据集的标准化混淆矩阵",
-    "xpack.ml.dataframe.analytics.classificationExploration.evaluateJobIdTitle": "分类作业 ID {jobId} 的评估",
     "xpack.ml.dataframe.analytics.classificationExploration.generalizationDocsCount": "{docsCount, plural, one {# 个文档} other {# 个文档}}已评估",
     "xpack.ml.dataframe.analytics.classificationExploration.showActions": "显示操作",
     "xpack.ml.dataframe.analytics.classificationExploration.showAllColumns": "显示所有列",
@@ -10769,13 +10768,11 @@
     "xpack.ml.dataframe.analytics.errorCallout.queryParsingErrorTitle": "无法解析查询。",
     "xpack.ml.dataframe.analytics.exploration.colorRangeLegendTitle": "功能影响分数",
     "xpack.ml.dataframe.analytics.explorationResults.documentsShownHelpText": "正在显示有相关预测存在的文档",
-    "xpack.ml.dataframe.analytics.explorationResults.fieldSelection": "已选择 {docFieldsCount, number} 个{docFieldsCount, plural, one {字段} other {字段}}中的 {selectedFieldsLength, number} 个",
     "xpack.ml.dataframe.analytics.explorationResults.firstDocumentsShownHelpText": "正在显示有相关预测存在的前 {searchSize} 个文档",
     "xpack.ml.dataframe.analytics.indexPatternPromptLinkText": "创建索引模式",
     "xpack.ml.dataframe.analytics.indexPatternPromptMessage": "不存在索引 {destIndex} 的索引模式。{destIndex} 的{linkToIndexPatternManagement}。",
     "xpack.ml.dataframe.analytics.jobCaps.errorTitle": "无法提取结果。加载索引的字段数据时发生错误。",
     "xpack.ml.dataframe.analytics.jobConfig.errorTitle": "无法提取结果。加载作业配置数据时发生错误。",
-    "xpack.ml.dataframe.analytics.regressionExploration.evaluateJobIdTitle": "回归作业 ID {jobId} 的评估",
     "xpack.ml.dataframe.analytics.regressionExploration.generalizationDocsCount": "{docsCount, plural, one {# 个文档} other {# 个文档}}已评估",
     "xpack.ml.dataframe.analytics.regressionExploration.generalizationErrorTitle": "泛化误差",
     "xpack.ml.dataframe.analytics.regressionExploration.generalizationFilterText": ".筛留训练数据。",
